Refactor: Add map panel.
