--- placeholder.js
@@ -1,3 +1,0 @@
-// Placeholder file to give input to rollup
-
-export default function () {}
