+++ src/i18n/index.js
@@ -1,4 +1,4 @@
-import {isDevelopmentOrTest} from '../utils/commonfunctions';
+import {isDevelopmentOrTest} from '../utils/commonFunctions';
 
 import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
