+++ src/swBuild.js
@@ -3,7 +3,7 @@ const workboxBuild = require('workbox-build');
 const buildSW = () => {
   workboxBuild
     .injectManifest({
-      swSrc: 'src/sw-template.js',
+      swSrc: 'src/swTemplate.js',
       swDest: 'build/service-worker.js',
       globDirectory: 'build',
       globPatterns: ['**/*.{html,woff2,js,css}'],
